Power supply issue and add serial console section
Changes to multiple sections to make the use of an independent
power supply mandatory.

Add a new section called Serial Console to describe how to use
wired connection to obtain information about the board's state.

Minor changes to other parts of the tutorial that remove outdated
content or make it consistent with the latest update.